WIP/MAJOR: mworker: change fork order on the master process
This is an experiment in which the worker is forked before loading its
configuration.

The "master" process will start without any configuration, only
configuring the master CLI. After the fork of the worker, the worker
will read the configuration and will behave like a normal non-mworker
haproxy in foreground mode.

This should allow to cleanup the whole mworker mode startup, removing
ugly hack like the "atexit()" for the configuration parser fallback.

Current patch should be able to start and reload, it doesn't support
reload failure yet.