MAJOR: SPOE: Mark the SPOE as unmaintained
From the 3.1, the SPOE is considered as unmaintained. No support must be
expected anymore on it on 3.1 and upper. It was not removed because there is
no alternative solution and some users rely on it. To avoid any definitive
breakage, we keep it. It was deprecated in 3.0 because we are not able to
make it evolve and it become impossible to maintain it. It is one more step
towards the SPOE suppression. Thus, the 3.0 is the last version with a true
SPOE support. Patchs applicable to newer versions will of course be
applied. But only as best effort.